requirements.txt
# --- CORE FRAMEWORK ---
Flask
flask-socketio
python-dotenv
requests
gunicorn

# --- SERVER ENGINE ---
Werkzeug

# --- SECURITY ENGINES ---
python-owasp-zap-v2.4
GitPython

# --- ARTIFICIAL INTELLIGENCE ---
groq
google-genai

# --- MACHINE LEARNING ---
# We use a broad range here to let Railway pick the best compatible version
torch>=2.2.0 --index-url https://download.pytorch.org/whl/cpu
transformers>=4.37.2
safetensors

# --- REPORTING ---
fpdf

# --- UTILITIES ---
colorama
urllib3
beautifulsoup4
flask-talisman
selenium
webdriver-manager